HACK: allow atttempting to fork/exec foreign PE
This also disables claiming to support x86_64 and i686 on arm64.  This
will probably break ARM64EC, and whatever FEX emulation plugin that may
or may not work for i686 (I can't find any concrete instructions on how
to use it, just a reference to a registry key to plug FEX into, if I
knew where to get the binary to put there).